Implement cached fetching and build script for fetching data pre-build
