+++ .env.example
@@ -1,3 +1,2 @@
-NODE_ENV=production
 PUBLIC_SITE_URL=http://localhost:3000/
 PUBLIC_BASEPATH=
