+++ package.json
@@ -5,7 +5,6 @@
   "scripts": {
     "dev": "astro dev",
     "build": "astro build",
-    "fetch:api": "tsx scripts/fetch-data.ts",
     "preview": "astro preview",
     "check": "astro check",
     "lint": "oxlint",
@@ -19,12 +18,10 @@
     "@radix-ui/react-visually-hidden": "^1.2.4",
     "@tanstack/react-store": "^0.11.0",
     "@tanstack/react-table": "^8.21.3",
-    "@tanstack/react-virtual": "^3.13.24",
     "@visx/group": "^3.12.0",
     "@visx/scale": "^3.12.0",
     "@visx/shape": "^3.12.0",
     "astro": "^6.3.1",
-    "babel-plugin-react-compiler": "^1.0.0",
     "clsx": "^2.1.1",
     "cva": "1.0.0-beta.4",
     "fuse.js": "^7.3.0",
@@ -37,16 +34,13 @@
     "use-debounce": "^10.1.1"
   },
   "devDependencies": {
-    "@tailwindcss/postcss": "^4.3.0",
     "@tailwindcss/vite": "^4.3.0",
     "@types/node": "^25.7.0",
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
     "oxfmt": "^0.49.0",
     "oxlint": "^1.64.0",
-    "postcss": "^8.5.14",
     "tailwindcss": "^4.3.0",
-    "tsx": "^4.22.0",
     "typescript": "^6.0.3"
   },
   "packageManager": "pnpm@11.1.0",
